Move detailed page documentation to "more"
